* don't try to ping poller while polling -- it won't reply
git-svn-id: https://svn.code.sf.net/p/penguintv/code/trunk@815 a0c6dda6-8710-0410-9408-9ac0bdba4c84

+++ penguintv/penguintv.py
@@ -433,6 +433,7 @@ class PenguinTVApp(gobject.GObject):
 		self._remote_poller_pid = -2
 		rundir = os.path.split(utils.__file__)[0]
 		if rundir == "": rundir = "./"
+		#logging.debug("RUN POLLER NOW")
 		logging.debug("running poller: %s %s" % ('/usr/bin/env python', os.path.join(rundir, 'Poller.py')))
 		subprocess.Popen(['/usr/bin/env', 'python', 
 			os.path.join(rundir, 'Poller.py')])
@@ -518,18 +519,21 @@ class PenguinTVApp(gobject.GObject):
 				self._remote_poller_pid = 0
 				self._spawn_poller()
 				return True
-			try:
-				#and is it still responding?
-				self._remote_poller.ping()
-				logging.debug("Poller still running")
-			except Exception, e:
-				logging.debug("trying to reget poller")
-				os.kill(self._remote_poller_pid, 15)
-				self._spawn_poller()
-				if self._get_poller():
-					logging.debug("restarted poller")
-				else:
-					logging.debug("failed to restart poller")
+			if self._polling_taskinfo == -1:
+				try:
+					#and is it still responding?
+					self._remote_poller.ping()
+					logging.debug("Poller still running")
+				except Exception, e:
+					logging.debug("no ping response, rerunning poller: %s" % str(e))
+					os.kill(self._remote_poller_pid, 15)
+					self._spawn_poller()
+					if self._get_poller():
+						logging.debug("restarted poller")
+					else:
+						logging.debug("failed to restart poller")
+			else:
+				logging.debug("not pinging while polling")
 				
 		return True
